refactor: move updateLastIncidentData function
